dfile: add a cast to int to the returned NULL value

+++ src/dos/dfile.c
@@ -155,7 +155,7 @@ int al_findfirst(AL_CONST char *pattern, struct al_ffblk *info, int attrib)
 
    if (!ff_data) {
       *allegro_errno = ENOMEM;
-      return NULL;
+      return (int)NULL;
    }
 
    /* attach it to the info structure */
